pyproject.toml
[tool.poetry]
name = "kms"
version = "0.1.0"
description = ""
authors = ["Lincride <admin@lincride.ng>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.1"
djangorestframework-api-key = "^3.0.0"
djangorestframework = "^3.15.2"
markdown = "^3.7"
django-filter = "^24.3"
gunicorn = "^23.0.0"
daphne = "^4.1.2"
dj-database-url = "^2.2.0"
django-cors-headers = "^4.4.0"
django-storages = {extras = ["s3"], version = "^1.14.4"}
hvac = "^2.3.0"
sentry-sdk = "^2.17.0"
celery = "^5.4.0"
redis = "^5.1.1"
django-redis = "^5.4.0"
flower = "^2.0.1"
cuid2 = "^2.0.1"
django-jazzmin = "^3.0.1"
django-debug-toolbar = "^4.4.6"
pyotp = "^2.9.0"
djangorestframework-simplejwt = "^5.3.1"
email-validator = "^2.2.0"
django-import-export = "^4.1.1"
channels-redis = "^4.2.0"
channels = "^4.1.0"
whitenoise = "^6.7.0"
django-celery-results = "^2.5.1"
cryptography = "^43.0.3"
drf-spectacular = "^0.27.2"
drf-standardized-errors = "^0.14.1"
psycopg2-binary = "^2.9.10"
django-celery-beat = "^2.7.0"
colorlog = "^6.9.0"
polars = "^1.16.0"
pandas = "^2.2.3"
pycryptodome = "^3.21.0"
django-extensions = "^4.1"
polyline = "^2.0.4"
aiokafka = "^0.12.0"
python-decouple = "^3.8"
uvicorn = {extras = ["standard"], version = "^0.38.0"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
